+++ package.json
@@ -13,6 +13,7 @@
     "postcss-cli": "^9.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.1.2"
   },
@@ -20,8 +21,7 @@
     "start": "craco start",
     "build": "craco build",
     "test": "craco test",
-    "eject": "react-scripts eject",
-    "watch:css": "postcss -w src/css/tailwind.css -o src/css/reusable.css"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
