+++ src/components/footer.jsx
@@ -1,7 +1,7 @@
 function Header() {
   return (
-    <div className="absolute bottom-0 bg-green-300 w-full py-3">
-      <p className="font-sans font-bold text-black text-lg text-center mx-4">Created by Reggie Gunawan - 2021</p>
+    <div className="absolute bottom-0 bg-transparent w-full py-3">
+      <p className="font-sans text-white text-base text-center mx-4">Created by Reggie Gunawan</p>
     </div>
   );
 }
