+++ src/pages/main/index.jsx
@@ -1,0 +1,14 @@
+import Home from "../../components/Home";
+import Header from "../../components/Header";
+import Footer from "../../components/Footer";
+import CardList from "../pcp/index.jsx";
+
+function Main() {
+  return (
+    <div className="min-h-screen flex flex-col">
+      <Home />
+    </div>
+  );
+}
+
+export default Main;
